+++ src/util.hpp
@@ -7,7 +7,7 @@
 
 SDL_Surface *LoadImage(const char filename[],bool trans=false);
 
-void ApplySurface(SDL_Surface *src,SDL_Surface *dst,int x,int y);
+void ApplySurface(SDL_Surface *src,SDL_Surface *dst,short x,short y);
 
 void ApplySurface(SDL_Surface *src,SDL_Surface *dst,SDL_Rect offset);
